Improve virtual machine performance
This commit improves the performance of a basic MUXLEQ virtual machine,
especially by leveraging branch prediction.

It is assumed that most operations are not I/O, and the compiler will
optimize branch prediction for common case. For the SUBLEQ operations,
the majority of the time, the branch (r == 0 || r & SZ) will not happen,
leading to faster execution of the usual case.

By using c & MASK, memory accesses are limited to the size of the VM
memory, avoiding undefined behavior when out-of-bounds access occurs.